+++ noodles-csi/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.58.0 - 2026-07-24
 
 ### Changed